feat: accept resource links to native external references in v10 type generator (#407)
Trying to use `cf-content-types-generator` with a Contentful schema that
includes fields with native external references currently results in an
exception and failure to generate types. This PR updates
`renderPropResourceLinkV10` to recognize the `ExternalResourceLink` type
and allow it. When encountering an unknown `allowedResources` value it
logs the value and treats it as `EntryFieldTypes.ExternalResourceLink`.
Includes updated tests.